Add static_cast to prevent build error with clang compiler when using the C++11 standard. This error is caused by a narrowing conversion of a non-constant expression from unsigned int to int

+++ src/Infrastructure/Mesh/src/ESMCI_MBMesh_Regrid_Glue.C
@@ -314,7 +314,7 @@ void MBMesh_regrid_create(void **meshsrcpp, ESMCI::Array **arraysrcpp, ESMCI::Po
     std::pair<UInt,UInt> iisize = wts.count_matrix_entries();
     int num_entries = iisize.first;
     int *iientries = new int[2*iisize.first];
-    int larg[2] = {2, iisize.first};
+    int larg[2] = {2, static_cast<int>(iisize.first)};
     // Gather the list
     ESMCI::InterArray<int> ii(iientries, 2, larg);
     ESMCI::InterArray<int> *iiptr = &ii;
